Dual Threaded Stake Miner
The stake miner runs two independent threads each operating on a separate
set of inputs. Every thread has a dedicated cache.

There is a new -stakefence option in hours. Inputs aged under this limit
are processed by the first staking thread. Those above the boundary are
a resposibility of the second staking thread. It is possible to run the
stake miner in a single threaded mode by setting this option to zero.

The -reservebalance option has been removed.